Add discount order creation

+++ db/migrate/20200228193939_create_discounts.rb
@@ -2,7 +2,7 @@ class CreateDiscounts < ActiveRecord::Migration[5.1]
   def change
     create_table :discounts do |t|
       t.integer :threshold
-      t.integer :percent
+      t.float :percent
     end
   end
 end
